fix: Use only declared zmq constants for monitor events (#20)
Checking "EVENT_" prefix is not a very accurate implementation,
as the main (unreleased) branch of pyzmq shows:
https://github.com/zeromq/pyzmq/blob/8d406b6e1/zmq/constants.py#L271

Though, previously we anyway hard-coded only "EVENT_"-prefixed constants
and there is no difference in terms of the accuracy.

I personally want to avoid using `dir()` to iterate over arbitrary
module contents to *import* constant declarations, but this is
inevitable because pyzmq seems to heavily refactor the constant
modules to better support static typing in the upcoming release.

- `zmq.constants` will become a separate module based on Python enums
  instead of being imported from Cython-based runtime-loaded
  `zmq.sugar.constants`.
- Currently mypy does not recognize `zmq.sugar.constants.base_names`
  because it's an imported name.
- `zmq.utils.constant_names.base_names` will be removed (!) and this
  means that it's a private API of the pyzmq library.
- It is the best to use the next release's `zmq.constants.Event` enum
  declaration, but it is not released yet.